fix(android): unify HA badge and entity-sheet icon/color mapping
Android had two divergent Home Assistant visual mappings: the on-video
badge used `badgeVisual`/`edgeOn` (a faithful port of the desktop
`haVisualFor`, matching the other clients) while the entity sheet and
more-info dialog used a separate `haVisual` with a different palette
(open door #FFC107 vs #FFB143), a different active-set (unlocked/playing/
active treated as active), and different glyphs (motion -> Sensors vs
figure.run). The same entity read differently between the badge and the
sheet, and a badge tap even rendered the badge one way but its detail
dialog another.

Collapse both onto one canonical source of truth. Extract the badge
mapping (palette, `BadgeVisual`, `edgeOn`, `defaultVisual`, `badgeVisual`,
`parseHexColor`) into a shared `HaVisual.kt` that both the badge overlay
and the sheet/dialog call, and delete the divergent `haVisual`. Fold the
sheet's extra active-state tokens (opening/unlocked/playing/active, plus
locked -> off) into the one `edgeOn` so those entities keep reading as
active under the unified map instead of falling back to indeterminate.

The icon vocabulary is deliberately left as the badge's existing set;
the closed-vocabulary rework is tracked separately (#438).

Closes #437
