+++ apps/android/app/src/main/java/video/crumb/app/feature/live/HaBadgeOverlay.kt
@@ -19,25 +19,6 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Bolt
-import androidx.compose.material.icons.filled.DirectionsRun
-import androidx.compose.material.icons.filled.Doorbell
-import androidx.compose.material.icons.filled.Garage
-import androidx.compose.material.icons.filled.Lightbulb
-import androidx.compose.material.icons.filled.LocalFireDepartment
-import androidx.compose.material.icons.filled.Lock
-import androidx.compose.material.icons.filled.MovieFilter
-import androidx.compose.material.icons.filled.Person
-import androidx.compose.material.icons.filled.Pets
-import androidx.compose.material.icons.filled.Power
-import androidx.compose.material.icons.filled.PowerOff
-import androidx.compose.material.icons.filled.SensorDoor
-import androidx.compose.material.icons.filled.SensorWindow
-import androidx.compose.material.icons.filled.Sensors
-import androidx.compose.material.icons.filled.Thermostat
-import androidx.compose.material.icons.filled.Videocam
-import androidx.compose.material.icons.filled.WaterDrop
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
@@ -54,7 +35,6 @@ import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.luminance
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.text.font.FontWeight
@@ -66,7 +46,6 @@ import video.crumb.app.data.HaStatesResponse
 import java.time.Duration
 import java.time.Instant
 import java.time.OffsetDateTime
-import java.util.Locale
 import kotlin.math.min
 
 // On-video Home Assistant badge layer for the live fullscreen screen — the
@@ -79,139 +58,13 @@ import kotlin.math.min
 // `edgeOn` EXACTLY (including the honesty rule: unknown/stale never reads as a
 // confident "off/closed").
 
-// Palette — matches desktop `ha_icons.dart`.
-private val BadgeGrey = Color(0xFF8E8E93)
-private val BadgeAmber = Color(0xFFFFB143) // open (door/window/garage)
-private val BadgeNeutral = Color(0xFFB9C2CC) // closed/off but KNOWN — not grey
-private val BadgeBlue = Color(0xFF33C3FF) // motion / occupancy active
-private val BadgeGreen = Color(0xFF2BA84A) // switch on
-private val BadgeWarmYellow = Color(0xFFFFCC33) // light on
+// The canonical entity visual mapping (`badgeVisual`, `BadgeVisual`, the palette,
+// `parseHexColor`) lives in `HaVisual.kt`, shared with the entity sheet + more-
+// info dialog so an entity reads identically wherever Android draws it (#437).
 private val BadgeDefaultBg = Color(0xFF17171B) // near-black opaque chip
 
 private const val BASE_REF_PX = 22f // reference badge size at pane-scale 1.0
 private const val REF_SHORT_SIDE = 320f
-
-/** Resolved badge look: state color, glyph, and a friendly state label. */
-data class BadgeVisual(val color: Color, val icon: ImageVector, val label: String)
-
-/**
- * HA `state` -> on / off / indeterminate, mirroring desktop `edgeOn` (and the
- * server `edge_on`) EXACTLY. `null` = indeterminate (unavailable/unknown/
- * anything else) and is NEVER treated as off.
- */
-private fun edgeOn(state: String): Boolean? = when (state.trim().lowercase(Locale.US)) {
-    "on", "open", "detected", "true", "home", "motion", "occupied" -> true
-    "off", "closed", "clear", "false", "not_home", "no_motion" -> false
-    else -> null
-}
-
-/** device_class -> Crumb label slug (mirrors desktop `labelForDeviceClass`). */
-private fun labelForDeviceClass(deviceClass: String?): String =
-    when (deviceClass?.trim()?.lowercase(Locale.US)) {
-        "motion", "moving", "vibration" -> "motion"
-        "occupancy", "presence" -> "occupancy"
-        "door", "opening" -> "door"
-        "window" -> "window"
-        "garage_door" -> "garage"
-        else -> "sensor"
-    }
-
-/** Operator-pickable per-badge icon override (migration 0059 slug -> glyph). */
-private val badgeIconSlugs: Map<String, ImageVector> = mapOf(
-    "door" to Icons.Filled.SensorDoor,
-    "window" to Icons.Filled.SensorWindow,
-    "garage" to Icons.Filled.Garage,
-    "motion" to Icons.Filled.DirectionsRun,
-    "person" to Icons.Filled.Person,
-    "lightbulb" to Icons.Filled.Lightbulb,
-    "power" to Icons.Filled.Power,
-    "switch" to Icons.Filled.Power,
-    "lock" to Icons.Filled.Lock,
-    "doorbell" to Icons.Filled.Doorbell,
-    "water" to Icons.Filled.WaterDrop,
-    "leak" to Icons.Filled.WaterDrop,
-    "fire" to Icons.Filled.LocalFireDepartment,
-    "smoke" to Icons.Filled.LocalFireDepartment,
-    "thermostat" to Icons.Filled.Thermostat,
-    "camera" to Icons.Filled.Videocam,
-    "pet" to Icons.Filled.Pets,
-    "scene" to Icons.Filled.MovieFilter,
-    "sensor" to Icons.Filled.Sensors,
-    "energy" to Icons.Filled.Bolt,
-)
-
-private fun defaultIcon(domain: String, deviceClass: String?): ImageVector = when {
-    domain == "light" -> Icons.Filled.Lightbulb
-    domain == "switch" -> Icons.Filled.Power
-    domain == "scene" -> Icons.Filled.MovieFilter
-    else -> when (labelForDeviceClass(deviceClass)) {
-        "door" -> Icons.Filled.SensorDoor
-        "window" -> Icons.Filled.SensorWindow
-        "garage" -> Icons.Filled.Garage
-        "motion" -> Icons.Filled.DirectionsRun
-        "occupancy" -> Icons.Filled.Person
-        else -> Icons.Filled.Sensors
-    }
-}
-
-/** Port of desktop `_haVisualDefault` — device-class/domain + state -> look. */
-private fun defaultVisual(
-    domain: String,
-    deviceClass: String?,
-    state: String?,
-    stale: Boolean,
-): BadgeVisual {
-    if (domain == "scene") return BadgeVisual(BadgeNeutral, Icons.Filled.MovieFilter, "Scene")
-    val on = if (state == null || stale) null else edgeOn(state)
-    if (on == null) {
-        return BadgeVisual(BadgeGrey.copy(alpha = 0.6f), defaultIcon(domain, deviceClass), state ?: "Unknown")
-    }
-    if (domain == "light") {
-        return BadgeVisual(if (on) BadgeWarmYellow else BadgeGrey, Icons.Filled.Lightbulb, if (on) "On" else "Off")
-    }
-    if (domain == "switch") {
-        return BadgeVisual(
-            if (on) BadgeGreen else BadgeGrey,
-            if (on) Icons.Filled.Power else Icons.Filled.PowerOff,
-            if (on) "On" else "Off",
-        )
-    }
-    return when (labelForDeviceClass(deviceClass)) {
-        "door" -> BadgeVisual(if (on) BadgeAmber else BadgeNeutral, Icons.Filled.SensorDoor, if (on) "Open" else "Closed")
-        "window" -> BadgeVisual(if (on) BadgeAmber else BadgeNeutral, Icons.Filled.SensorWindow, if (on) "Open" else "Closed")
-        "garage" -> BadgeVisual(if (on) BadgeAmber else BadgeNeutral, Icons.Filled.Garage, if (on) "Open" else "Closed")
-        "motion" -> BadgeVisual(if (on) BadgeBlue else BadgeGrey, Icons.Filled.DirectionsRun, if (on) "Motion" else "Clear")
-        "occupancy" -> BadgeVisual(if (on) BadgeBlue else BadgeGrey, Icons.Filled.Person, if (on) "Occupied" else "Clear")
-        else -> BadgeVisual(if (on) BadgeBlue else BadgeGrey, Icons.Filled.Sensors, if (on) "Active" else "Clear")
-    }
-}
-
-/**
- * Port of desktop `haVisualFor`: the default look, then the operator's per-badge
- * icon/color overrides. The color override applies ONLY to a KNOWN reading
- * (active full-strength, inactive dimmed) — never to unknown/stale, where the
- * grey honesty treatment must win.
- */
-private fun badgeVisual(link: HaLinkDto, state: String?, stale: Boolean): BadgeVisual {
-    val base = defaultVisual(link.domain, link.deviceClass, state, stale)
-    val overrideIcon = link.overlayIcon?.let { badgeIconSlugs[it] }
-    val on = if (state == null || stale || link.domain == "scene") null else edgeOn(state)
-    val colorOverride = parseHexColor(link.overlayColor)
-    val color = if (colorOverride != null && on != null) {
-        if (on) colorOverride else colorOverride.copy(alpha = 0.45f)
-    } else {
-        base.color
-    }
-    return BadgeVisual(color, overrideIcon ?: base.icon, base.label)
-}
-
-/** Parse `#RRGGBB` -> opaque [Color], or null if absent/malformed. */
-private fun parseHexColor(hex: String?): Color? {
-    val h = hex?.trim()?.removePrefix("#") ?: return null
-    if (h.length != 6) return null
-    val v = h.toLongOrNull(16) ?: return null
-    return Color(0xFF000000L or v)
-}
 
 /** Compact "just now / 5s / 3m / 2h / 4d" from an RFC3339 timestamp. */
 private fun relativeAgo(iso: String?): String? {
